Deprecates mocha path in benefit of testing
